requirements.txt
boto3==1.6.16
Django==2.0.3
djangorestframework==3.7.7
django-storages==1.6.5
dj-database-url==0.5.0
gunicorn==19.7.1
Pillow==5.0.0
psycopg2==2.7.4
